third_party/WebKit/LayoutTests/fast/dom/shadow/get-destination-insertion-points-skips-user-agent-shadow.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
</head>
<body>
<script>
description("Tests that getDestinationInsertionPoints() should skip insertion points in user-agent shadow roots.");

document.body.appendChild(
    createDOM('div', {},
              createDOM('div', {'id': 'host'},
                        createShadowRoot(
                            createDOM('detail', {'id': 'detail'},
                                      createDOM('div', {'id': 'detail-child'}),
                                      createDOM('content', {'id': 'content'}))),
                        createDOM('div', {'id': 'host-child'}))));

shouldBeEqualAsArray(document.getElementById('host-child').getDestinationInsertionPoints(),
                     [getNodeInComposedTree('host/content')]);
shouldBeZero("getNodeInComposedTree('host/detail-child').getDestinationInsertionPoints().length");
</script>
</body>
</html>
